fixed sidebar sizing
